ZQS-814 Move flip_wavefunction and flip_amplitudes functions to zquantum.core.wavefunction
