+++ src/python/zquantum/core/symbolic_simulator.py
@@ -4,9 +4,9 @@ import numpy as np
 from sympy import Symbol
 from zquantum.core.circuits import Circuit
 from zquantum.core.circuits.layouts import CircuitConnectivity
-from zquantum.core.interfaces.backend import QuantumSimulator, flip_wavefunction
+from zquantum.core.interfaces.backend import QuantumSimulator
 from zquantum.core.measurement import Measurements, sample_from_wavefunction
-from zquantum.core.wavefunction import Wavefunction
+from zquantum.core.wavefunction import Wavefunction, flip_wavefunction
 
 
 class SymbolicSimulator(QuantumSimulator):
